ADD: Trashの描画, 機能実装

+++ Assets/Scripts/Domain/IPresenter/ITrashRenderer.cs
@@ -2,6 +2,6 @@ namespace Domain.IPresenter
 {
     public interface ITrashRenderer
     {
-        void UpdateTrashRemain(float remain);
+        void UpdateTrashRemain(int remain);
     }
 }
